fix tests by adding sampler and sensibility
Several of the tests relied on randomisation, but R 3.6 fixed a sampling
bug that made set.seed irreproducible unless the user explicitly changed
the RNG kind. This is basically a patch for that.

I also changed the expectation for the print functions since the results
were dependent on the width of the R terminal when run.